+++ include/cinolib/meshes/abstract_mesh.h
@@ -133,7 +133,7 @@ class AbstractMesh
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
         virtual uint verts_per_poly(const uint pid) const = 0;
-        virtual uint edges_per_poly(const uint pid) const { return this->p2e.at(pid).size(); }
+        virtual uint edges_per_poly(const uint pid) const { return uint(this->p2e.at(pid).size()); }
 
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
