+++ include/cinolib/meshes/abstract_polygonmesh.h
@@ -96,7 +96,7 @@ class AbstractPolygonMesh : public AbstractMesh<M,V,E,P>
 
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
-        uint verts_per_poly(const uint pid) const override { return this->polys.at(pid).size(); }
+        uint verts_per_poly(const uint pid) const override { return uint(this->polys.at(pid).size()); }
 
         //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
